Move test_builder to evm.tools
